refactor: migrate package structure from org.simplecommerce.ai.commerce to org.sc.ai.cli
BREAKING CHANGE: Package structure has been migrated using OpenRewrite recipe

- Migrated all Java classes from org.simplecommerce.ai.commerce.* to org.sc.ai.cli.*
- Updated group ID in build.gradle from org.simplecommerce.ai to org.sc.ai
- Updated OpenRewrite dependencies to latest versions
- Updated native image metadata references
- Updated Gradle documentation generation tasks

Package mappings:
- org.simplecommerce.ai.commerce → org.sc.ai.cli
- org.simplecommerce.ai.commerce.chat → org.sc.ai.cli.chat
- org.simplecommerce.ai.commerce.command → org.sc.ai.cli.command
- org.simplecommerce.ai.commerce.config → org.sc.ai.cli.config
- org.simplecommerce.ai.commerce.rag → org.sc.ai.cli.rag

This change simplifies the package naming convention and aligns better with the
project's CLI-focused purpose while maintaining all existing functionality.
