+++ .github/copilot-instructions.md
@@ -5,3 +5,4 @@ applyTo: '**'
 # Git instructions
 
 - Prefer conventional commits for commit messages.
+- For longer PR descriptions, use the `create_pull_request` tool
