+++ build.gradle
@@ -54,8 +54,8 @@ dependencies {
     testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
     
     // OpenRewrite dependencies
-    rewrite 'org.openrewrite.recipe:rewrite-java:8.45.4'
-    rewrite 'org.openrewrite.recipe:rewrite-spring:5.24.0'
+    rewrite 'org.openrewrite:rewrite-java:8.56.0'
+    rewrite 'org.openrewrite.recipe:rewrite-spring:6.9.0'
 }
 
 dependencyManagement {
